+++ framework/model/src/main/java/is/codion/framework/model/AbstractEntityTableModel.java
@@ -44,7 +44,7 @@ import java.util.function.Consumer;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 
-import static is.codion.framework.model.EntityTableConditionModel.entityTableConditionModel;
+import static is.codion.framework.model.EntityConditionModel.entityConditionModel;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.toList;
@@ -70,7 +70,7 @@ public abstract class AbstractEntityTableModel<E extends EntityEditModel> implem
 	 * @param tableModel the filter model
 	 */
 	protected AbstractEntityTableModel(E editModel, FilterModel<Entity> tableModel) {
-		this(editModel, tableModel, new DefaultEntityQueryModel(entityTableConditionModel(editModel.entityType(), editModel.connectionProvider())));
+		this(editModel, tableModel, new DefaultEntityQueryModel(entityConditionModel(editModel.entityType(), editModel.connectionProvider())));
 	}
 
 	/**